refactor: update JPA repositories to support Long IDs

+++ server/src/main/java/com/xtheggx/monedaDOS/repository/AlertaRepository.java
@@ -3,6 +3,6 @@ package com.xtheggx.monedaDOS.repository;
 import com.xtheggx.monedaDOS.model.Alerta;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlertaRepository extends JpaRepository <Alerta, Integer>{
+public interface AlertaRepository extends JpaRepository <Alerta, Long>{
 
 }
