+++ server/src/main/java/com/xtheggx/monedaDOS/repository/CategoriaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CategoriaRepository extends JpaRepository<Categoria, Integer> {
+public interface CategoriaRepository extends JpaRepository<Categoria, Long> {
     List<Categoria> findByTipo(CategoriaTipo tipo);
 
     List<Categoria> findByUsuarioIdUsuario(Long usuarioIdUsuario);
