+++ server/src/main/java/com/xtheggx/monedaDOS/repository/ConfiguracionUsuarioRepository.java
@@ -2,6 +2,6 @@ package com.xtheggx.monedaDOS.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ConfiguracionUsuarioRepository extends JpaRepository<com.xtheggx.monedaDOS.model.ConfiguracionUsuario, Integer> {
+public interface ConfiguracionUsuarioRepository extends JpaRepository<com.xtheggx.monedaDOS.model.ConfiguracionUsuario, Long> {
 
 }
